+++ client/src/redux/reducers/datafiles.reducers.js
@@ -2,25 +2,69 @@ export const initialSystemState = {
   defaultHost: '',
   systemList: [],
   error: false,
-  errorMessage: null
+  errorMessage: null,
+  loading: false,
+  definitions: []
+};
+
+export const addSystemDefinition = (system, definitionList) => {
+  return [
+    ...definitionList.filter(existing => existing.id !== system.id),
+    system
+  ];
 };
 
 export function systems(state = initialSystemState, action) {
   switch (action.type) {
+    case 'FETCH_SYSTEMS_STARTED':
+      return {
+        ...state,
+        error: false,
+        errorMessage: null,
+        loading: true
+      };
     case 'FETCH_SYSTEMS_SUCCESS':
       return {
         ...state,
         systemList: action.payload.system_list,
-        defaultHost: action.payload.default_host
+        defaultHost: action.payload.default_host,
+        loading: false
       };
     case 'FETCH_SYSTEMS_ERROR':
-      return { ...state, error: true, errorMessage: action.payload };
+      return {
+        ...state,
+        error: true,
+        errorMessage: action.payload,
+        loading: false
+      };
+    case 'FETCH_SYSTEM_DEFINITION_STARTED':
+      return {
+        ...state,
+        error: false,
+        errorMessage: null,
+        loading: true
+      };
+    case 'FETCH_SYSTEM_DEFINITION_SUCCESS':
+      return {
+        ...state,
+        definitions: addSystemDefinition(action.payload, state.definitions),
+        error: false,
+        errorMessage: null,
+        loading: false
+      };
+    case 'FETCH_SYSTEM_DEFINITION_ERROR':
+      return {
+        ...state,
+        error: true,
+        errorMessage: action.payload,
+        loading: false
+      };
     default:
       return state;
   }
 }
 
-const initialFilesState = {
+export const initialFilesState = {
   loading: {
     FilesListing: false,
     modal: false
@@ -31,7 +75,13 @@ const initialFilesState = {
     copy: {},
     select: {},
     upload: {},
-    trash: {}
+    trash: {},
+    link: {
+      method: null,
+      url: '',
+      error: null,
+      loading: false
+    }
   },
   loadingScroll: {
     FilesListing: false,
@@ -59,7 +109,12 @@ const initialFilesState = {
     FilesListing: true,
     modal: true
   },
+  nextPageToken: {
+    FilesListing: null,
+    modal: null
+  },
   modals: {
+    addproject: false,
     preview: false,
     move: false,
     copy: false,
@@ -67,8 +122,11 @@ const initialFilesState = {
     upload: false,
     mkdir: false,
     rename: false,
+    link: false,
     pushKeys: false,
-    trash: false
+    trash: false,
+    manageproject: false,
+    editproject: false
   },
   modalProps: {
     preview: {},
@@ -78,9 +136,15 @@ const initialFilesState = {
     upload: {},
     mkdir: {},
     rename: {},
-    pushKeys: {}
+    pushKeys: {},
+    link: {},
+    showpath: {}
   },
-  previewHref: ''
+  preview: {
+    href: '',
+    content: '',
+    isLoading: true
+  }
 };
 
 let selectedSet, enabled, setValue;
@@ -104,6 +168,10 @@ export function files(state = initialFilesState, action) {
         selectAll: {
           ...state.selectAll,
           [action.payload.section]: false
+        },
+        nextPageToken: {
+          ...state.nextPageToken,
+          [action.payload.section]: null
         }
       };
     case 'FETCH_FILES_SUCCESS':
@@ -118,6 +186,10 @@ export function files(state = initialFilesState, action) {
         reachedEnd: {
           ...state.reachedEnd,
           [action.payload.section]: action.payload.reachedEnd
+        },
+        nextPageToken: {
+          ...state.nextPageToken,
+          [action.payload.section]: action.payload.nextPageToken
         }
       };
     case 'FETCH_FILES_ERROR':
@@ -159,6 +231,10 @@ export function files(state = initialFilesState, action) {
         reachedEnd: {
           ...state.reachedEnd,
           [action.payload.section]: action.payload.reachedEnd
+        },
+        nextPageToken: {
+          ...state.nextPageToken,
+          [action.payload.section]: action.payload.nextPageToken
         }
       };
     case 'SCROLL_FILES_ERR':
@@ -239,12 +315,15 @@ export function files(state = initialFilesState, action) {
           }
         }
       };
-    case 'DATA_FILES_SET_PREVIEW_HREF':
+    case 'DATA_FILES_SET_PREVIEW_CONTENT':
       return {
         ...state,
-        previewHref: action.payload.href
+        preview: {
+          href: action.payload.href,
+          content: action.payload.content,
+          isLoading: action.payload.isLoading
+        }
       };
-
     case 'DATA_FILES_TOGGLE_MODAL':
       return {
         ...state,
@@ -255,6 +334,31 @@ export function files(state = initialFilesState, action) {
         modalProps: {
           ...state.modalProps,
           [action.payload.operation]: action.payload.props
+        }
+      };
+    case 'DATA_FILES_SET_MODAL_PROPS':
+      return {
+        ...state,
+        modalProps: {
+          ...state.modalProps,
+          [action.payload.operation]: action.payload.props
+        }
+      };
+    case 'DATA_FILES_CLEAR_PROJECT_SELECTION':
+      return {
+        ...state,
+        error: {
+          ...state.error,
+          FilesListing: false
+        },
+        params: {
+          ...state.params,
+          FilesListing: {
+            api: 'tapis',
+            scheme: 'projects',
+            system: '',
+            path: ''
+          }
         }
       };
     default: